Module kaleidoscope_parser::parser

The module defining functions to parse a given Kaleidoscope input. This parser can be used to generate Abstract Syntax Trees (ASTs), from which LLVM IR code can be created.

Structs

Parser

The parser struct that converts a Kaleidoscope program into an Abstract Syntax Tree.

ParserToken

The token manager struct that determines when a Token should be released for the parser’s use or when a new token should be read from the stream. This struct is marked private because the functionality of this struct is only useful when applied in conjunction with Parser’s methods.

Type Definitions

ParseResult

The return type of most parser functions in Parser.