Struct kaleidoscope_parser::parser::Parser

pub struct Parser {
    current_token: ParserToken,
}

The parser struct that converts a Kaleidoscope program into an Abstract Syntax Tree.

Fields
current_token: ParserToken
Implementations

impl Parser

pub fn new() -> Self

Create a new parser.

pub fn next_token(&mut self, token: Token) -> Result<&mut Self>

Replace the token the parser is currently reading with another one.

pub fn replace_used_token(&mut self, token: Token) -> Result<&mut Self>

Replace the token the parser is currently reading with another one, given that it has been used or there was no token in the first place.

pub(crate) fn mark_unused(&mut self) -> &mut Self

Mark the token in the manager as unused.

pub(crate) fn mark_used(&mut self) -> &mut Self

Mark the token in the manager as used.

fn grab_token_from_tokenizer<'a, 'b: 'a>(
&mut self,
$crate::tokenizer::LexerTupleMut: LexerTupleMut<'a, 'b>
) -> Result<&mut Self>

Replace the token with a new token from the stream and tokenizer.

fn grab_if_used<'a, 'b: 'a>(
&mut self,
$crate::tokenizer::LexerTupleMut: LexerTupleMut<'a, 'b>
) -> Result<&mut Self>

Pull a new token from the tokenizer and file stream if the token stored by the parser has been used.

fn get_current_token(&mut self) -> Option<Token>

Get the current token being stored, marking as used in the process.

pub fn peek_current_token(&self) -> Option<Token>

Peek at the current token being stored, without marking it as used.

fn find_matching_right_round_bracket<'a, 'b: 'a>(
&mut self,
lbracket_index: FileIndex,
$crate::tokenizer::LexerTupleMut: LexerTupleMut<'a, 'b>
) -> Result<Token>

Helper code that finds a right bracket that cancels out a left bracket.

pub fn parse_comma_expression_list<'a, 'b: 'a>(
&mut self,
$crate::tokenizer::LexerTupleMut: LexerTupleMut<'a, 'b>,
left_bracket: Bracket
) -> Result<Option<Vec<Box<dyn ExprNode>>>>

Parse a program that contains a list of expressions. Each expression in the list is separated by a comma token (i.e. “,”) and the entire list is flanked by a left bracket before the first expression and its corresponding right bracket after the last expression.

pub fn parse_top_level_expression<'a, 'b: 'a>(
&mut self,
$crate::tokenizer::LexerTupleMut: LexerTupleMut<'a, 'b>
) -> ParseResult<dyn ExprNode>

Converts an expression into an anonymous function.

pub fn parse_expression<'a, 'b: 'a>(
&mut self,
$crate::tokenizer::LexerTupleMut: LexerTupleMut<'a, 'b>
) -> ParseResult<dyn ExprNode>

Parse an expression.

pub fn parse_primary_expression<'a, 'b: 'a>(
&mut self,
$crate::tokenizer::LexerTupleMut: LexerTupleMut<'a, 'b>
) -> ParseResult<dyn ExprNode>

Parse a so-called “primary” expression. You can think of primary expressions as simple expressions, which means that parsing it is relatively trivial. This category includes integers, floats, variables and other expressions wrapped around two corresponding round brackets.

pub fn parse_integer_expression<'a, 'b: 'a>(
&mut self,
$crate::tokenizer::LexerTupleMut: LexerTupleMut<'a, 'b>
) -> ParseResult<dyn ExprNode>

Parse an integer expression.

pub fn parse_float_expression<'a, 'b: 'a>(
&mut self,
$crate::tokenizer::LexerTupleMut: LexerTupleMut<'a, 'b>
) -> ParseResult<dyn ExprNode>

Parse a float expression.

pub fn parse_variable_expression<'a, 'b: 'a>(
&mut self,
$crate::tokenizer::LexerTupleMut: LexerTupleMut<'a, 'b>
) -> ParseResult<dyn ExprNode>

Parse a variable expression.

pub fn parse_round_bracket_expression<'a, 'b: 'a>(
&mut self,
$crate::tokenizer::LexerTupleMut: LexerTupleMut<'a, 'b>
) -> ParseResult<dyn ExprNode>

Parse an expression wrapped inside 2 round brackets.

pub fn parse_binary_operator_rhs_expression<'a, 'b: 'a>(
&mut self,
lhs: Box<dyn ExprNode>,
loperator: Operator,
minimum_operator_precedence: BinaryOperatorPrecedence,
escaped_from_inner: &mut bool,
depth: usize,
$crate::tokenizer::LexerTupleMut: LexerTupleMut<'a, 'b>
) -> ParseResult<dyn ExprNode>

Parse a binary operator expression. This is similar to simple math equations like 1 + 1 or 5 * 3.

pub fn parse_function_call_expression<'a, 'b: 'a>(
&mut self,
$crate::tokenizer::LexerTupleMut: LexerTupleMut<'a, 'b>
) -> ParseResult<dyn ExprNode>

Parse a call to a function. Such an expression comes in two parts, the first is the name of the function and the second are the arguments passed to the function. In Kaleidoscope, the syntax of function calls are similar to those found in C or any of its descendant languages.

For example, a function call like pow(4, 2) would be converted into a [FunctionCallNode] where the function being called is pow and the arguments (in the correct order) are 4 and 2.

As the Parser is not allowed to reverse tokens when reading a stream, this function has to shoulder the burden of also doing the job of Parser::parse_variable_expression. If it encounters an identifier which does not have a left, round bracket token (“(”) following it, it will do what Parser::parse_variable_expression does and returns a [VariableExpressionNode].

pub fn parse_function_prototype<'a, 'b: 'a>(
&mut self,
$crate::tokenizer::LexerTupleMut: LexerTupleMut<'a, 'b>
) -> ParseResult<FunctionPrototypeNode>

Parse a function prototype.

pub fn parse_function<'a, 'b: 'a>(
&mut self,
$crate::tokenizer::LexerTupleMut: LexerTupleMut<'a, 'b>
) -> ParseResult<FunctionNode>

Parse a function definition.

pub fn parse_extern_function<'a, 'b: 'a>(
&mut self,
$crate::tokenizer::LexerTupleMut: LexerTupleMut<'a, 'b>
) -> ParseResult<ExternFunctionNode>

Parse an extern function declaration.